requirements.txt
alembic==1.10.2
dnspython==2.3.0
email_validator==1.3.1
Flask==2.2.3
flask_bcrypt==1.0.1
flask_login==0.6.2
flask_migrate==4.0.4
flask-moment-1.0.5
flask_sqlalchemy==3.0.3
flask_wtf==1.1.1
pymysql==1.0.2
SQLAlchemy==2.0.7
WTForms==3.0.1
